A lot of refactoring!
