+++ src/components/search/Facets.vue
@@ -1,32 +1,46 @@
 <template>
-	<div class="facet-container">
-		<div>
-			<Transition
-				name="fade"
-				mode="out-in"
+	<div
+		class="search-facets active"
+		ref="facetsContainer"
+	>
+		<div class="facet-background">
+			<button
+				class="facet-close-button"
+				@click="searchResultStore.toggleShowFacets(false)"
 			>
+				<span class="material-icons">close</span>
+			</button>
+
+			<div class="facet-container">
 				<div>
-					<h2 class="headline">{{ t('search.channels') }}</h2>
-					<div
-						class="checkbox"
-						v-for="(singleFacet, index) in currentFacetNr as unknown as FacetPair[]"
-						:key="index + 'facet'"
+					<Transition
+						name="fade"
+						mode="out-in"
 					>
-						<kb-checkboxcomponent
-							:fqkey="'creator_affiliation'"
-							:title="channelFacets[index]?.title"
-							:number="channelFacets[index]?.number"
-							:value="filterExists('creator_affiliation', channelFacets[index]?.title)"
-							:inslide="index"
-							:content="!searchResultStore.loading"
-						/>
-					</div>
+						<div>
+							<h2 class="headline">{{ t('search.channels') }}</h2>
+							<div
+								class="checkbox"
+								v-for="(singleFacet, index) in currentFacetNr as unknown as FacetPair[]"
+								:key="index + 'facet'"
+							>
+								<kb-checkboxcomponent
+									:fqkey="'creator_affiliation'"
+									:title="channelFacets[index]?.title"
+									:number="channelFacets[index]?.number"
+									:value="filterExists('creator_affiliation', channelFacets[index]?.title)"
+									:inslide="index"
+									:content="!searchResultStore.loading"
+								/>
+							</div>
+						</div>
+					</Transition>
+					<CategoryTags
+						:categories="categoryFacets"
+						:category-nr="categoryNr"
+					></CategoryTags>
 				</div>
-			</Transition>
-			<CategoryTags
-				:categories="categoryFacets"
-				:category-nr="categoryNr"
-			></CategoryTags>
+			</div>
 		</div>
 	</div>
 </template>
@@ -53,7 +67,7 @@ export default defineComponent({
 		facetResults: { type: Object as PropType<FacetResultType>, required: true },
 	},
 
-	setup(props, { emit }) {
+	setup(props) {
 		const searchResultStore = useSearchResultStore();
 		const showFacets = ref(false);
 		const currentFacets = ref(Object as unknown as FacetResultType);
@@ -61,6 +75,8 @@ export default defineComponent({
 		const channelFacets = ref([] as FacetPair[]);
 		const categoryFacets = ref([] as FacetPair[]);
 		const categoryNr = ref(0);
+
+		const facetsContainer = ref<HTMLElement | null>(null);
 
 		const lastUpdate = ref(0);
 		const route = useRoute();
@@ -86,7 +102,6 @@ export default defineComponent({
 						channelFacets.value = [] as FacetPair[];
 						categoryFacets.value = [] as FacetPair[];
 
-						showFacets.value = false;
 						currentFacets.value = newFacets;
 						channelFacets.value = simplifyFacets(newFacets['creator_affiliation']);
 						categoryFacets.value = simplifyFacets(newFacets['categories']);
@@ -103,15 +118,50 @@ export default defineComponent({
 			window.removeEventListener('filter-update', filterUpdateHelper);
 		});
 
+		/* This is because Vue3 composition API has this weird bug that when a ref is wrapped in a v-if
+		   the ref is not actually present in onMounted, which is super weird. But we can watch for when it enters.
+		   https://github.com/vuejs/composition-api/issues/296
+		*/
+		watch(
+			() => facetsContainer.value,
+			(newValue: HTMLElement | null, oldValue: HTMLElement | null) => {
+				if (facetsContainer.value && newValue !== oldValue) {
+					window.innerWidth > 800
+						? searchResultStore.toggleShowFacets(true)
+						: searchResultStore.toggleShowFacets(false);
+				}
+			},
+		);
+
 		const updateFilters = (e: CustomEvent) => {
 			const routeQueries = e.detail.add ? addFilter(route, e.detail.filter) : removeFilter(route, e.detail.filter);
 			routeQueries.start = 0;
 			router.push({ query: routeQueries });
 		};
 
+		watch(
+			() => searchResultStore.showFacets,
+			() => {
+				toggleFacets();
+			},
+		);
+
+		const toggleFacets = () => {
+			if (searchResultStore.showFacets) {
+				facetsContainer.value?.classList.add('active');
+				//resultContainer.value?.classList.add('fullwidth');
+
+				window.document.body.classList.add('remove-body-scroll');
+			} else {
+				facetsContainer.value?.classList.remove('active');
+				//resultContainer.value?.classList.remove('fullwidth');
+				window.document.body.classList.remove('remove-body-scroll');
+			}
+		};
+
 		const filterUpdateHelper = (e: Event) => {
 			updateFilters(e as CustomEvent);
-			emit('facetUpdate');
+			window.innerWidth < 800 ? searchResultStore.toggleShowFacets(false) : null;
 		};
 
 		return {
@@ -130,6 +180,8 @@ export default defineComponent({
 			addFilter,
 			removeFilter,
 			route,
+			toggleFacets,
+			facetsContainer,
 			t,
 		};
 	},
@@ -143,6 +195,13 @@ export default defineComponent({
 	transition: all 1.3s linear;
 	height: auto;
 	flex-direction: column;
+}
+
+.search-facets.active {
+	visibility: visible;
+	pointer-events: all;
+	transform: translateX(0%);
+	width: 100%;
 }
 
 h2 {
@@ -161,9 +220,78 @@ h2 {
 	padding: 3px 0px;
 }
 
+.facet-close-button {
+	border: 0px;
+	background-color: transparent;
+	font-size: 0px;
+	position: absolute;
+	top: 10px;
+	z-index: 25;
+	right: 10px;
+	cursor: pointer;
+}
+.search-facets {
+	transition: all 0.25s cubic-bezier(0.455, 0.03, 0.515, 0.955) 0s;
+	min-width: 0px;
+	margin-bottom: 20px;
+	position: fixed;
+	top: 0px;
+	left: 0px;
+	z-index: 5;
+	background-color: white;
+	overflow-y: scroll;
+	height: 100vh;
+	visibility: hidden;
+	pointer-events: none;
+	box-sizing: border-box;
+	padding: 40px 15% 0px 15%;
+	transform: translateX(-100%);
+}
+
+@media (min-width: 640px) {
+	.search-facets {
+		top: 20%;
+		height: 60%;
+		box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
+		width: 70%;
+	}
+
+	.search-facets::-webkit-scrollbar-track {
+		background-color: transparent;
+	}
+
+	.search-facets.active {
+		width: 70%;
+	}
+}
 @media (min-width: 800px) {
 	.facet-container {
 		flex-direction: column;
+	}
+
+	.facet-background {
+		background-color: rgba(30, 30, 30, 0.1);
+	}
+	.facet-close-button {
+		display: none;
+	}
+	.search-facets {
+		position: initial;
+		background-color: initial;
+		width: 0px;
+		margin-right: 0px;
+		overflow-y: initial;
+		overflow-x: hidden;
+		padding: 0px 0px;
+		transform: translateX(0%);
+		opacity: 0;
+		box-shadow: initial;
+	}
+	.search-facets.active {
+		width: 290px;
+		margin-right: 30px;
+		min-width: 300px;
+		opacity: 1;
 	}
 }
 </style>
